accounts

+++ server-src/api/user.js
@@ -48,12 +48,10 @@ const methods = {
   },
   
   deleteUser: function ({ user_id }) {
-    return db.collection('filters').deleteMany(
-      { user_id: user_id }
-    ).then(r => {
-      return db.collection('users').deleteOne(
-        { _id: user_id }
-      )
+    return db.collection('filters').deleteMany({ user_id: user_id }).then(r => {
+      return db.collection('accounts').deleteMany({ user_id: user_id })
+    }).then(r => {
+      return db.collection('users').deleteOne({ _id: user_id })
     })
   },
